Housekeeping: separate utils, name types with T
