+++ src/types/regional-data.ts
@@ -1,6 +1,6 @@
-import { Province, Municipality } from "./region";
+import { TProvince, TMunicipality } from "./region";
 
-type ProvinceKeys = keyof typeof Province;
+type ProvinceKeys = keyof typeof TProvince;
 
 export type TempRegionalData = Partial<Record<ProvinceKeys, string[]>>;
-export type RegionalData = Partial<Record<ProvinceKeys, Municipality>>;
+export type RegionalData = Partial<Record<ProvinceKeys, TMunicipality>>;
